refactor(design-token-panel): split color-tweak panel into tabbed shell
Move the monolithic src/components/color-tweak-panel.tsx (~1135 lines)
into src/components/design-token-tweak/ as a tabbed panel (Spacing,
Font, Size, Color). Only the Color tab has a body in this change — it
is a verbatim move of the existing Palette/Base/Semantic/Shiki sections.
The other three tabs render placeholder content; their real controls
land in Sub 2/3/4.

- Rename the exposed setting colorTweakPanel -> designTokenPanel and
  keep colorTweakPanel as a deprecated alias that still enables the
  panel (either key being truthy flips it on).
- Rename the toggle event to toggle-design-token-panel; the panel
  still listens for toggle-color-tweak-panel as an alias.
- Add localStorage v1 (zudo-doc-tweak-state) -> v2
  (zudo-doc-tweak-state-v2) migration that namespaces the Color
  sub-state under state.color so future tabs can add siblings. v2
  wins when both keys are present; malformed JSON falls back safely
  with console.warn.
- Grow the panel footprint: width min(1200px, 80vw), height
  min(800px, 80vh) on desktop. Below 900px the panel switches to
  width min(100vw - 16px, 500px), is centered, and disables drag.
- Tab bar uses role=tablist/tab/tabpanel with aria-selected,
  aria-controls, roving tabindex, and Left/Right/Home/End keyboard
  navigation. Default tab is Color.
- Vitest coverage for the migration (valid v1, partial v1, corrupt v1,
  v1+v2 both with v2 winning, v2 malformed falling back to v1) and
  the deprecated-alias gate.

Refs #346, closes #347
