+++ src/components/design-token-tweak/controls/.gitkeep
@@ -1,0 +1,2 @@
+Placeholder for shared control primitives (slider, preview tile, palette selector, etc.)
+used by the Spacing / Font / Size / Color tabs. Populated in Sub 2 / 3 / 4.
