+++ src/components/design-token-tweak/state/persist.ts
@@ -1,0 +1,44 @@
+import { useCallback } from "react";
+import {
+  type TweakState,
+  type ColorTweakState,
+  applyFullState,
+  savePersistedState,
+} from "./tweak-state";
+
+type SetState<T> = (updater: (prev: T | null) => T | null) => void;
+
+/**
+ * Hook returning a `persist` callback and tab-scoped helpers.
+ *
+ * `persist(updater)` runs the whole-state updater, applies the resulting CSS
+ * variables to `:root`, and writes the v2 key to localStorage. Tab components
+ * use the more focused `persistColor(updater)` helper so they can stay
+ * unaware of the v2 envelope.
+ */
+export function usePersist(setState: SetState<TweakState>) {
+  const persist = useCallback(
+    (updater: (prev: TweakState) => TweakState) => {
+      setState((prev) => {
+        if (!prev) return prev;
+        const next = updater(prev);
+        applyFullState(next);
+        savePersistedState(next);
+        return next;
+      });
+    },
+    [setState],
+  );
+
+  const persistColor = useCallback(
+    (updater: (prev: ColorTweakState) => ColorTweakState) => {
+      persist((prev) => ({ ...prev, color: updater(prev.color) }));
+    },
+    [persist],
+  );
+
+  return { persist, persistColor };
+}
+
+export type Persist = ReturnType<typeof usePersist>["persist"];
+export type PersistColor = ReturnType<typeof usePersist>["persistColor"];
