+++ src/layouts/doc-layout.astro
@@ -21,13 +21,17 @@ import VersionSwitcher from "@/components/version-switcher.astro";
 import VersionBanner from "@/components/version-banner.astro";
 import { getVersionAvailability } from "@/utils/version-availability";
 import Footer from "@/components/footer.astro";
-import ColorTweakPanel from "@/components/color-tweak-panel";
+import DesignTokenTweakPanel from "@/components/design-token-tweak";
 import AiChatModal from "@/components/ai-chat-modal";
 import MockInit from "@/components/mock-init";
 import DesktopSidebarToggle from "@/components/desktop-sidebar-toggle";
 import FindInPageInit from "@/components/find-in-page-init";
 import ImageEnlarge from "@/components/image-enlarge";
 import { SEMANTIC_DEFAULTS, SEMANTIC_CSS_NAMES } from "@/config/color-scheme-utils";
+
+// Either the new `designTokenPanel` flag or the deprecated `colorTweakPanel`
+// alias enables the panel. Old projects keep working for one release.
+const designTokenPanelEnabled = Boolean(settings.designTokenPanel || settings.colorTweakPanel);
 
 interface Props {
   title: string;
@@ -93,29 +97,44 @@ const contentTransition = {
     <meta property="og:title" content={fullTitle} />
     {description && <meta property="og:description" content={description} />}
     <ColorSchemeProvider />
-    {settings.colorTweakPanel && (
+    {designTokenPanelEnabled && (
       <script is:inline define:vars={{ tweakSemanticDefaults: SEMANTIC_DEFAULTS, tweakSemanticCss: SEMANTIC_CSS_NAMES }}>
         (function () {
+          var V1_KEY = "zudo-doc-tweak-state";
+          var V2_KEY = "zudo-doc-tweak-state-v2";
+          function readColorState() {
+            // v2 wins; fall back to v1 shape if v2 is absent.
+            var rawV2 = null;
+            try { rawV2 = localStorage.getItem(V2_KEY); } catch (e) {}
+            if (rawV2) {
+              try {
+                var parsed = JSON.parse(rawV2);
+                if (parsed && parsed.color) return parsed.color;
+              } catch (e) { /* fall through */ }
+            }
+            var rawV1 = null;
+            try { rawV1 = localStorage.getItem(V1_KEY); } catch (e) {}
+            if (rawV1) {
+              try { return JSON.parse(rawV1); } catch (e) { /* malformed; ignore */ }
+            }
+            return null;
+          }
           function applyTweakState() {
-            var raw = localStorage.getItem("zudo-doc-tweak-state");
-            if (!raw) return;
-            try {
-              var s = JSON.parse(raw);
-              if (!s.palette || s.palette.length !== 16 || s.background === undefined) return;
-              var root = document.documentElement;
-              for (var i = 0; i < 16; i++) root.style.setProperty("--zd-" + i, s.palette[i]);
-              root.style.setProperty("--zd-bg", s.palette[s.background]);
-              root.style.setProperty("--zd-fg", s.palette[s.foreground]);
-              root.style.setProperty("--zd-cursor", s.palette[s.cursor]);
-              root.style.setProperty("--zd-sel-bg", s.palette[s.selectionBg]);
-              root.style.setProperty("--zd-sel-fg", s.palette[s.selectionFg]);
-              for (var key in tweakSemanticCss) {
-                var m = s.semanticMappings && s.semanticMappings[key];
-                if (m === undefined) m = tweakSemanticDefaults[key];
-                var val = m === "bg" ? s.palette[s.background] : m === "fg" ? s.palette[s.foreground] : s.palette[m];
-                root.style.setProperty(tweakSemanticCss[key], val);
-              }
-            } catch (e) { localStorage.removeItem("zudo-doc-tweak-state"); }
+            var s = readColorState();
+            if (!s || !s.palette || s.palette.length !== 16 || s.background === undefined) return;
+            var root = document.documentElement;
+            for (var i = 0; i < 16; i++) root.style.setProperty("--zd-" + i, s.palette[i]);
+            root.style.setProperty("--zd-bg", s.palette[s.background]);
+            root.style.setProperty("--zd-fg", s.palette[s.foreground]);
+            root.style.setProperty("--zd-cursor", s.palette[s.cursor]);
+            root.style.setProperty("--zd-sel-bg", s.palette[s.selectionBg]);
+            root.style.setProperty("--zd-sel-fg", s.palette[s.selectionFg]);
+            for (var key in tweakSemanticCss) {
+              var m = s.semanticMappings && s.semanticMappings[key];
+              if (m === undefined) m = tweakSemanticDefaults[key];
+              var val = m === "bg" ? s.palette[s.background] : m === "fg" ? s.palette[s.foreground] : s.palette[m];
+              root.style.setProperty(tweakSemanticCss[key], val);
+            }
           }
           applyTweakState();
           document.addEventListener("astro:after-swap", applyTweakState);
@@ -224,7 +243,7 @@ const contentTransition = {
       <Footer lang={lang} />
     </div>
     {settings.aiAssistant && <AiChatModal basePath={withBase("/")} client:load />}
-    {settings.colorTweakPanel && <ColorTweakPanel client:only="preact" />}
+    {designTokenPanelEnabled && <DesignTokenTweakPanel client:only="preact" />}
     <CodeBlockEnhancer />
     <TabsInit />
     {settings.mermaid && <MermaidInit />}
